+++ src/assign/learning_rate.rs
@@ -1,12 +1,12 @@
 /// Var Rewarding based on Learning Rate Rewardin gand Reason Side Rewarding
 use {
     super::{AssignStack, VarRewardIF},
-    crate::types::*,
+    crate::{
+        state::{SearchStrategy, State},
+        types::*,
+    },
     std::slice::Iter,
 };
-
-/// a pair of the start value and upper bound of var decay rate.
-pub const REWARD_DECAY_RANGE: (f64, f64) = (0.80, 0.97);
 
 impl VarRewardIF for AssignStack {
     #[inline]
@@ -14,14 +14,12 @@ impl VarRewardIF for AssignStack {
         self.var[vi].reward
     }
     fn initialize_reward(&mut self, iterator: Iter<'_, usize>) {
-        self.activity_decay = REWARD_DECAY_RANGE.0;
-        self.activity_decay_max = REWARD_DECAY_RANGE.1;
-        self.reward_step = (REWARD_DECAY_RANGE.1 - REWARD_DECAY_RANGE.0) / 10_000.0;
+        self.reward_step = (self.activity_decay_max - self.activity_decay).abs() / 10_000.0;
         // big bang initialization
-        let mut v = 0.5;
+        let mut v = 0.25;
         for vi in iterator {
             self.var[*vi].reward = v;
-            v *= 0.9;
+            v *= 0.99;
         }
     }
     fn clear_reward(&mut self, vi: VarId) {
@@ -40,9 +38,10 @@ impl VarRewardIF for AssignStack {
         let v = &mut self.var[vi];
         let duration = (self.ordinal + 1 - v.timestamp) as f64;
         let decay = self.activity_decay;
+        let _decay = (1.0 - duration.ln() * (1.0 - self.activity_decay)).max(0.0);
         let rate = v.participated as f64 / duration;
         v.reward *= decay;
-        v.reward += (1.0 - decay) * rate;
+        v.reward += (1.0 - decay) * rate.sqrt();
         v.participated = 0;
     }
     fn reward_update(&mut self) {
@@ -50,5 +49,19 @@ impl VarRewardIF for AssignStack {
         self.activity_decay = self
             .activity_decay_max
             .min(self.activity_decay + self.reward_step);
+        // self.activity_decay = 1.0 - 1.0 / (1.0 + 0.5 * (self.num_restart as f64)).sqrt();
+    }
+    fn adjust_reward(&mut self, state: &State) {
+        if state.strategy.1 == self.num_conflict {
+            match state.strategy.0 {
+                SearchStrategy::LowDecisions => {
+                    self.activity_decay_max -= 0.02;
+                }
+                SearchStrategy::HighSuccesive => {
+                    self.activity_decay_max = (self.activity_decay_max + 0.005).min(0.999);
+                }
+                _ => (),
+            };
+        }
     }
 }