+++ src/assign/propagate.rs
@@ -1,7 +1,7 @@
 /// implement boolean constraint propagation, backjump
 /// This version can handle Chronological and Non Chronological Backtrack.
 use {
-    super::{AssignIF, AssignStack, VarHeapIF, VarRewardIF},
+    super::{AssignIF, AssignStack, VarHeapIF, VarRewardIF, VarSelectIF},
     crate::{
         cdb::{ClauseDBIF, WatchDBIF},
         types::*,
@@ -36,8 +36,6 @@ pub trait PropagateIF {
     fn propagate<C>(&mut self, cdb: &mut C) -> ClauseId
     where
         C: ClauseDBIF;
-    /// reset or copy phase data.
-    fn reset_assign_record(&mut self, flag: Flag, from: Option<Flag>);
 }
 
 macro_rules! var_assign {
@@ -148,6 +146,10 @@ impl PropagateIF for AssignStack {
     fn cancel_until(&mut self, lv: DecisionLevel) {
         if self.trail_lim.len() as u32 <= lv {
             return;
+        }
+        if self.best_assign {
+            self.save_phases();
+            self.best_assign = false;
         }
         let lim = self.trail_lim[lv as usize];
         let mut shift = lim;
@@ -274,70 +276,16 @@ impl PropagateIF for AssignStack {
             }
         }
         let na = self.trail.len() + self.num_eliminated_vars;
-        // if self.num_target_assign < na {
-        //     self.target_assign = true;
-        //     self.num_target_assign = na;
-        //     self.save_phase(Flag::TARGET_PHASE, self.build_target_at + 1 == self.num_propagation);
-        //     self.build_target_at = self.num_propagation;
-        // }
-        if self.num_best_assign < na {
+        if 0 < self.decision_level() && self.num_best_assign < na {
             self.best_assign = true;
             self.num_best_assign = na;
-            self.save_phase(
-                Flag::BEST_PHASE,
-                self.build_best_at + 1 == self.num_propagation,
-            );
-            self.build_best_at = self.num_propagation;
         }
         ClauseId::default()
-    }
-    fn reset_assign_record(&mut self, flag: Flag, from: Option<Flag>) {
-        match flag {
-            Flag::PHASE => {
-                if let Some(source) = from {
-                    for v in self.var.iter_mut().skip(1) {
-                        v.set(flag, v.is(source));
-                    }
-                }
-            }
-            Flag::BEST_PHASE => {
-                if let Some(source) = from {
-                    for v in self.var.iter_mut().skip(1) {
-                        v.set(flag, v.is(source));
-                    }
-                    self.build_best_at = self.num_propagation;
-                } else {
-                    self.num_best_assign = 0;
-                }
-            }
-            Flag::TARGET_PHASE => {
-                self.num_target_assign = self.num_eliminated_vars;
-                if let Some(source) = from {
-                    for v in self.var.iter_mut().skip(1) {
-                        v.set(flag, v.is(source));
-                    }
-                }
-            }
-            _ => {
-                #[cfg(feature = "boundary_check")]
-                panic!("invalid flag for reset_assign_record");
-            }
-        }
     }
 }
 
 impl AssignStack {
     fn level_up(&mut self) {
         self.trail_lim.push(self.trail.len());
-    }
-    fn save_phase(&mut self, flag: Flag, incremental: bool) {
-        let to = if incremental && 0 < self.decision_level() {
-            self.len_upto(self.decision_level() - 1)
-        } else {
-            0
-        };
-        for l in self.trail.iter().skip(to) {
-            self.var[l.vi()].set(flag, bool::from(*l));
-        }
     }
 }